select online or in person
